+++ packages/@expo/cli/src/start/server/middleware/__tests__/InterstitialPageMiddleware-test.ts
@@ -1,5 +1,5 @@
 import { getConfig, getNameFromConfig } from '@expo/config';
-import { getRuntimeVersionNullable } from '@expo/config-plugins/build/utils/Updates';
+import { getRuntimeVersionNullableAsync } from '@expo/config-plugins/build/utils/Updates';
 import { vol } from 'memfs';
 
 import { asMock } from '../../../../__tests__/asMock';
@@ -19,7 +19,7 @@ jest.mock('@expo/config', () => ({
   getNameFromConfig: jest.fn(() => 'my-app'),
 }));
 jest.mock('@expo/config-plugins/build/utils/Updates', () => ({
-  getRuntimeVersionNullable: jest.fn(),
+  getRuntimeVersionNullableAsync: jest.fn(),
 }));
 
 const asReq = (req: Partial<ServerRequest>) => req as ServerRequest;
@@ -56,14 +56,14 @@ describe('_shouldHandleRequest', () => {
   });
 });
 
-describe('_getProjectOptions', () => {
+describe('_getProjectOptionsAsync', () => {
   it('returns the project settings from the config', async () => {
     asMock(getNameFromConfig).mockReturnValueOnce({ appName: 'my-app' });
-    asMock(getRuntimeVersionNullable).mockReturnValueOnce('123');
+    asMock(getRuntimeVersionNullableAsync).mockResolvedValueOnce('123');
 
     const middleware = new InterstitialPageMiddleware('/');
 
-    expect(middleware._getProjectOptions('ios')).toEqual({
+    expect(await middleware._getProjectOptionsAsync('ios')).toEqual({
       appName: 'my-app',
       projectVersion: {
         type: 'runtime',
@@ -71,18 +71,19 @@ describe('_getProjectOptions', () => {
       },
     });
     expect(getConfig).toBeCalled();
-    expect(getRuntimeVersionNullable).toBeCalledWith(
+    expect(getRuntimeVersionNullableAsync).toBeCalledWith(
+      '/',
       { name: 'my-app', sdkVersion: '45.0.0', slug: 'my-app' },
       'ios'
     );
   });
   it('returns the project settings from the config with SDK version', async () => {
     asMock(getNameFromConfig).mockReturnValueOnce({ appName: 'my-app' });
-    asMock(getRuntimeVersionNullable).mockReturnValueOnce(null);
+    asMock(getRuntimeVersionNullableAsync).mockResolvedValueOnce(null);
 
     const middleware = new InterstitialPageMiddleware('/');
 
-    expect(middleware._getProjectOptions('ios')).toEqual({
+    expect(await middleware._getProjectOptionsAsync('ios')).toEqual({
       appName: 'my-app',
       projectVersion: {
         type: 'sdk',
@@ -90,7 +91,8 @@ describe('_getProjectOptions', () => {
       },
     });
     expect(getConfig).toBeCalled();
-    expect(getRuntimeVersionNullable).toBeCalledWith(
+    expect(getRuntimeVersionNullableAsync).toBeCalledWith(
+      '/',
       { name: 'my-app', sdkVersion: '45.0.0', slug: 'my-app' },
       'ios'
     );
@@ -147,13 +149,15 @@ describe('handleRequestAsync', () => {
   it('returns the interstitial page with platform header', async () => {
     const middleware = new InterstitialPageMiddleware('/');
 
-    middleware._getProjectOptions = jest.fn(() => ({
-      appName: 'App',
-      projectVersion: {
-        type: 'runtime',
-        version: '123',
-      },
-    }));
+    middleware._getProjectOptionsAsync = jest.fn(() =>
+      Promise.resolve({
+        appName: 'App',
+        projectVersion: {
+          type: 'runtime',
+          version: '123',
+        },
+      })
+    );
 
     middleware._getPageAsync = jest.fn(async () => 'mock-value');
 
@@ -182,13 +186,15 @@ describe('handleRequestAsync', () => {
   it('returns the interstitial page with user-agent header', async () => {
     const middleware = new InterstitialPageMiddleware('/');
 
-    middleware._getProjectOptions = jest.fn(() => ({
-      appName: 'App',
-      projectVersion: {
-        type: 'runtime',
-        version: '123',
-      },
-    }));
+    middleware._getProjectOptionsAsync = jest.fn(() =>
+      Promise.resolve({
+        appName: 'App',
+        projectVersion: {
+          type: 'runtime',
+          version: '123',
+        },
+      })
+    );
 
     middleware._getPageAsync = jest.fn(async () => 'mock-value');