+++ packages/@expo/cli/CHANGELOG.md
@@ -16,6 +16,7 @@
 ### 💡 Others
 
 - Refactor e2e tests.
+- Forward the project root path to `getRuntimeVersionAsync` and `getRuntimeVersionNullableAsync`. ([#24126](https://github.com/expo/expo/pull/24126) by [@mccraveiro](https://github.com/mccraveiro))
 
 ## 0.10.12 — 2023-09-11